Updated Test4
Added a second version of multiplier for doubles. This function object could be replaced with a template class but for the purposes of this portfolio I have included it separately for syntax examples. Makes uses of template display function for doubles. Creates a second multiplier object, then uses std::transform to iterate through the double's vector and multiply by the multiplier's set variable.